+++ Fowler.Net/FirstRun/Price.cs
@@ -4,7 +4,7 @@
     {
         public abstract Movie.PriceCodes GetPriceCode();
 
-        public abstract double GetCharge(int daysRented);
+        public abstract decimal GetCharge(int daysRented);
 
         public virtual int GetFrequentRenterPoints(int daysRented)
         {
@@ -16,11 +16,11 @@
     {
         public override Movie.PriceCodes GetPriceCode() => Movie.PriceCodes.Childrens;
 
-        public override double GetCharge(int daysRented)
+        public override decimal GetCharge(int daysRented)
         {
-            var result = 1.5;
+            var result = 1.5m;
             if (daysRented > 3)
-                result += (daysRented - 3) * 1.5;
+                result += (daysRented - 3) * 1.5m;
             return result;
         }
     }
@@ -28,9 +28,9 @@
     internal class NewReleasePrice : Price
     {
         public override Movie.PriceCodes GetPriceCode() => Movie.PriceCodes.NewRelease;
-        public override double GetCharge(int daysRented)
+        public override decimal GetCharge(int daysRented)
         {
-            return daysRented * 3.0;
+            return daysRented * 3.0m;
         }
 
         public override int GetFrequentRenterPoints(int daysRented)
@@ -42,11 +42,11 @@
     internal class RegularPrice : Price
     {
         public override Movie.PriceCodes GetPriceCode() => Movie.PriceCodes.Regular;
-        public override double GetCharge(int daysRented)
+        public override decimal GetCharge(int daysRented)
         {
-            var result = 2.0;
+            var result = 2.0m;
             if (daysRented > 2)
-                result += (daysRented - 2) * 1.5;
+                result += (daysRented - 2) * 1.5m;
             return result;
         }
     }
